feat: handling exceptions with GlobalHandlerException

+++ src/main/java/com/example/jparestful/service/CategoryServiceImpl.java
@@ -1,10 +1,13 @@
 package com.example.jparestful.service;
 
 import com.example.jparestful.dto.CategoryDTO;
+import com.example.jparestful.exception.DuplicateResourceException;
+import com.example.jparestful.exception.ResourceNotFoundException;
 import com.example.jparestful.mapper.CategoryMapper;
 import com.example.jparestful.model.Category;
 import com.example.jparestful.repository.CategoryRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -32,8 +35,17 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category create(CategoryDTO categoryDTO) {
-        Category category = CategoryMapper.toEntity(categoryDTO);
-        return categoryRepository.save(category);
+        if(categoryRepository.findByName(categoryDTO.getName()).isPresent()){
+            throw new DuplicateResourceException("Category", "name",  categoryDTO.getName());
+        }
+
+        try {
+            Category category = CategoryMapper.toEntity(categoryDTO);
+            return categoryRepository.save(category);
+        } catch (DataIntegrityViolationException e) {
+            throw new DuplicateResourceException("Category", "name",  categoryDTO.getName());
+        }
+
     }
 
     @Override
@@ -45,16 +57,25 @@ public class CategoryServiceImpl implements CategoryService {
     public Category update(UUID id, CategoryDTO categoryDTO) {
         return categoryRepository.findById(id)
                 .map(existing -> {
+                    categoryRepository.findByName(categoryDTO.getName())
+                                    .filter(category -> !category.getId().equals(id))
+                                            .ifPresent(category -> {
+                                                throw new DuplicateResourceException("Category", "name",  categoryDTO.getName());
+                                            });
+
                     existing.setName(categoryDTO.getName());
                     existing.setDescription(categoryDTO.getDescription());
                     return categoryRepository.save(existing);
                 })
-                .orElseThrow(() -> new RuntimeException("Category not found with id " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Category", id.toString()));
     }
 
 
     @Override
     public void deleteById(UUID id) {
+        if(!categoryRepository.existsById(id)){
+            throw new ResourceNotFoundException("Category", id.toString());
+        }
         categoryRepository.deleteById(id);
     }
 }
